# Make page require faster