--- a/test/isolated/require-page.test.js
+++ b/test/isolated/require-page.test.js
@@ -1,7 +1,7 @@
 /* global describe, it, expect */
 
 import { join } from 'path'
-import {getPagePath, normalizePagePath, pageNotFoundError} from '../../dist/server/require'
+import requirePage, {getPagePath, normalizePagePath, pageNotFoundError} from '../../dist/server/require'
 
 const dir = '/path/to/some/project'
 const dist = '.next'
@@ -53,5 +53,30 @@ describe('getPagePath', () => {
 
   it('Should throw with paths containing ../', () => {
     expect(() => getPagePath('/../../package.json', {dir, dist})).toThrow()
+  })
+})
+
+describe('requirePage', () => {
+  it('Should require /index.js when using /', () => {
+    const page = requirePage('/', {dir: __dirname, dist: '_resolvedata'})
+    expect(page.test).toBe('hello')
+  })
+
+  it('Should require /index.js when using /index', () => {
+    const page = requirePage('/index', {dir: __dirname, dist: '_resolvedata'})
+    expect(page.test).toBe('hello')
+  })
+
+  it('Should require /world.js when using /world', () => {
+    const page = requirePage('/world', {dir: __dirname, dist: '_resolvedata'})
+    expect(page.test).toBe('world')
+  })
+
+  it('Should throw when using /../../test.js', () => {
+    expect(() => requirePage('/../../test.js', {dir: __dirname, dist: '_resolvedata'})).toThrow()
+  })
+
+  it('Should throw when using non existent pages like /non-existent.js', () => {
+    expect(() => requirePage('/non-existent.js', {dir: __dirname, dist: '_resolvedata'})).toThrow()
   })
 })