… test/isolated/_resolvedata/.gitignore
… test/isolated/_resolvedata/dist/bundles/pages/index.js
… test/isolated/_resolvedata/dist/bundles/pages/world.js