@arunoda 2018-02-12
@timneutkens the `resolve` logic you used doesn't work well inside Windows as you expected.
See:

![resolve](https://user-images.githubusercontent.com/50838/36115880-6a9ba80c-105a-11e8-8f35-fcb3cd51101c.png)

May be you can check whether the page stays inside the .next directory. I think with that, we don't need these kind of extra logic.
(We may already have working logic for that)