@timneutkens 2018-02-13
Added the console.error for those throws.