@timneutkens 2018-03-20
Logic was improved here: https://github.com/zeit/next.js/pull/3976